Crate tree_sitter_java

This crate provides a Java grammar for the tree-sitter parsing library.

Typically, you will use the language function to add this grammar to a tree-sitter Parser, and then use the parser to parse some code:

use tree_sitter::Parser;

let code = r#"
    class Test {
        int double(int x) {
            return x * 2;
        }
    }
"#;
let mut parser = Parser::new();
parser.set_language(tree_sitter_java::language()).expect("Error loading Java grammar");
let parsed = parser.parse(code, None);

Constants

GRAMMAR

The source of the Java tree-sitter grammar description.

HIGHLIGHT_QUERY

The syntax highlighting query for this language.

NODE_TYPES

The content of the node-types.json file for this grammar.

TAGGING_QUERY

The symbol tagging query for this language.

Functions

language

Returns the tree-sitter Language for this grammar.